+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/PlotTypes/Marker.cs
@@ -1,35 +1,67 @@
 ﻿namespace ScottPlotCookbook.Recipes.PlotTypes;
 
-internal class Marker : RecipePageBase
+internal class Marker : ICategory
 {
-    public override RecipePageDetails PageDetails => new()
-    {
-        Chapter = Chapter.PlotTypes,
-        PageName = "Marker",
-        PageDescription = "Custom Markers placed on the plot in coordinate space",
-    };
+    public string Chapter => "Plot Types";
+    public string CategoryName => "Marker";
+    public string CategoryDescription => "Markers can be placed on the plot in coordinate space.";
 
-    internal class Quickstart : RecipeTestBase
+    internal class MarkerQuickstart : RecipeBase
     {
         public override string Name => "Marker Quickstart";
-        public override string Description => "Markers with custom shapes, sizes, and styles can be placed anywhere in coordinate space.";
+        public override string Description => "Markers are symbols placed at a " +
+            "location in coordinate space. Their shape, size, and color can be customized.";
 
         [Test]
-        public override void Recipe()
+        public override void Execute()
         {
-            var colormap = new ScottPlot.Colormaps.Turbo();
-            MarkerShape[] markerShapes = Enum.GetValues<MarkerShape>().ToArray();
-            Random rand = new(0);
+            myPlot.Add.Signal(Generate.Sin());
+            myPlot.Add.Signal(Generate.Cos());
+
+            myPlot.Add.Marker(25, .5);
+            myPlot.Add.Marker(35, .6);
+            myPlot.Add.Marker(45, .7);
+        }
+    }
+
+    internal class MarkerShapes : RecipeBase
+    {
+        public override string Name => "Marker Shapes";
+        public override string Description => "Many marker shapes are available.";
+
+        [Test]
+        public override void Execute()
+        {
+            ScottPlot.Colormaps.Turbo colormap = new();
+
             for (int i = 0; i < 100; i++)
             {
-                myPlot.Add.Marker(
-                    x: rand.NextDouble(),
-                    y: rand.NextDouble(),
-                    size: 5 + rand.NextSingle() * 20,
-                    shape: markerShapes[rand.Next(markerShapes.Length)],
-                    color: colormap.GetColor(rand.NextDouble())
-                    );
+                MarkerShape shape = Generate.RandomMarkerShape();
+                Coordinates location = Generate.RandomLocation();
+                float size = Generate.RandomInteger(5, 10);
+                Color color = Generate.RandomColor(colormap);
+
+                myPlot.Add.Marker(location, shape, size, color);
             }
+        }
+    }
+
+    internal class MarkerLegend : RecipeBase
+    {
+        public override string Name => "Marker Legend";
+        public override string Description => "Markers with labels appear in the legend.";
+
+        [Test]
+        public override void Execute()
+        {
+            var sin = myPlot.Add.Signal(Generate.Sin());
+            var cos = myPlot.Add.Signal(Generate.Cos());
+            var marker = myPlot.Add.Marker(25, .5);
+
+            sin.Label = "Sine";
+            cos.Label = "Cosine";
+            marker.Label = "Marker";
+            myPlot.ShowLegend();
         }
     }
 }