+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Generate: Added methods for generating random marker shapes and colors
 * Generate: `Random()` is deprecated in favor of `RandomSample()`
 * Plot: Added `ShowLegend()` and `HideLegend()` helper methods which set `Plot.Legend.IsVisible`
+* Marker Plot: `Plot.AddMarker()` can now be used to place a single marker on the plot (#3076, #2806) _Thanks @Gray-lab_
 
 ## ScottPlot 4.1.70 (in development)
 * Population Plot: Improved performance for populations with curves that run off the screen (#3054) _Thanks @Em3a-c and @cornford_